修改bug：当数据文件不文件不存在时报错

+++ datarw/userrw.go
@@ -29,8 +29,9 @@ func GetUsers() []entity.User {
 func SaveUsers(usersToSave []entity.User) {
 	filePath := "datarw/Users.json"
 	//清空原文件
-	os.Truncate(filePath, 0)
-
+	if existFile(filePath) {
+		os.Truncate(filePath, 0)
+	}
 	//转为json串
 	josnStr, err := json.Marshal(usersToSave)
 	checkError(err)
@@ -64,7 +65,9 @@ func GetCurUser() *entity.User {
 func SaveCurUser(userToSave *entity.User) {
 	filePath := "datarw/CurUser.json"
 	//清空原文件
-	os.Truncate(filePath, 0)
+	if existFile(filePath) {
+		os.Truncate(filePath, 0)
+	}
 
 	if userToSave != nil {
 		//转为json串
